requirements.txt
python-telegram-bot[all]>=20.7
httpx
